Refactors image loading and display
Moves image loading and display logic to a dedicated view, improving code organization and reusability.

Enhances the user experience by adding a fade-in animation when the image loads.
Removes redundant card overlay from the Pokedex view, centralizing it within the dedicated view.
Replaces the AsyncImageView in the detail view with a standard Image view for a simpler implementation.