Refactors data loading to view models
Moves asynchronous data loading logic from the views' `task` modifiers into the view models' initializers.

This ensures that data loading begins as soon as the view model is initialized,
improving the perceived responsiveness of the UI.

The view models are also updated to conform to their protocols directly, and
the protocol conformance is moved into private extensions.