Refactors search initialization
Moves the initialization of the Pokemon source data in the search view model to the `onAppear` lifecycle method. This ensures the data is ready when the view appears, rather than relying on a `task` modifier that depends on the `pokemonData` which could lead to timing issues.

Also, uses keypath for whitespace check to improve code readability.