Improves sprite loading and sorting logic
Moves `ImageColorAnalyzer` to the `Misc` group for better organization.

Avoids redundant sprite loading by checking if the sprite is already loaded in `ItemSpriteViewModel` and `PokemonViewModel`.

Ensures that the Pokemon list is not unintentionally reset during sorting by safely unwrapping `self` within the sorting task. This prevents potential crashes if the view model is deallocated during the sorting process.